+++ Source/Krypton Components/Krypton.Toolkit/Values/ColorButtonValues.cs
@@ -1,4 +1,4 @@
-#region BSD License
+﻿#region BSD License
 /*
  * 
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
@@ -36,6 +36,7 @@ public class ColorButtonValues : Storage,
     private Color _emptyBorderColor;
     private Rectangle _selectedRect;
     private byte _roundedCorners;
+    private readonly OverlayImageValues _overlayImage;
     #endregion
 
     #region Events
@@ -62,6 +63,7 @@ public class ColorButtonValues : Storage,
         _extraText = _defaultExtraText;
         ImageStates = CreateImageStates();
         ImageStates.NeedPaint = needPaint;
+        _overlayImage = new OverlayImageValues(needPaint);
         _emptyBorderColor = Color.Gray;
         _selectedColor = Color.Red;
         _selectedRect = new Rectangle(0, 12, 16, 4);
@@ -81,6 +83,7 @@ public class ColorButtonValues : Storage,
                                       (Text == _defaultText) &&
                                       (ExtraText == _defaultExtraText)
                                       && (_roundedCorners == 0)
+                                      && _overlayImage.IsDefault
     ;
 
     #endregion
@@ -317,6 +320,19 @@ public class ColorButtonValues : Storage,
 
     #endregion
 
+    #region OverlayImage
+    /// <summary>
+    /// Gets access to the overlay image values.
+    /// </summary>
+    [Category(@"Visuals")]
+    [Description(@"Overlay image values.")]
+    [DesignerSerializationVisibility(DesignerSerializationVisibility.Content)]
+    public OverlayImageValues OverlayImage => _overlayImage;
+
+    private bool ShouldSerializeOverlayImage() => !_overlayImage.IsDefault;
+
+    #endregion
+
     #region IContentValues
     /// <summary>
     /// Gets the content image.
@@ -412,42 +428,42 @@ public class ColorButtonValues : Storage,
     /// </summary>
     /// <param name="state">The state for which the overlay image is needed.</param>
     /// <returns>Overlay image value, or null if no overlay image is set.</returns>
-    public virtual Image? GetOverlayImage(PaletteState state) => null;
+    public virtual Image? GetOverlayImage(PaletteState state) => _overlayImage.GetImage(state);
 
     /// <summary>
     /// Gets the overlay image color that should be transparent.
     /// </summary>
     /// <param name="state">The state for which the overlay image is needed.</param>
     /// <returns>Color value.</returns>
-    public virtual Color GetOverlayImageTransparentColor(PaletteState state) => SharedStaticVariables.EMPTY_COLOR;
+    public virtual Color GetOverlayImageTransparentColor(PaletteState state) => _overlayImage.ImageTransparentColor;
 
     /// <summary>
     /// Gets the position of the overlay image relative to the main image.
     /// </summary>
     /// <param name="state">The state for which the overlay position is needed.</param>
     /// <returns>Overlay image position.</returns>
-    public virtual OverlayImagePosition GetOverlayImagePosition(PaletteState state) => OverlayImagePosition.TopRight;
+    public virtual OverlayImagePosition GetOverlayImagePosition(PaletteState state) => _overlayImage.Position;
 
     /// <summary>
     /// Gets the scaling mode for the overlay image.
     /// </summary>
     /// <param name="state">The state for which the overlay scale mode is needed.</param>
     /// <returns>Overlay image scale mode.</returns>
-    public virtual OverlayImageScaleMode GetOverlayImageScaleMode(PaletteState state) => OverlayImageScaleMode.None;
+    public virtual OverlayImageScaleMode GetOverlayImageScaleMode(PaletteState state) => _overlayImage.ScaleMode;
 
     /// <summary>
     /// Gets the scale factor for the overlay image (used when scale mode is Percentage or ProportionalToMain).
     /// </summary>
     /// <param name="state">The state for which the overlay scale factor is needed.</param>
     /// <returns>Scale factor (0.0 to 2.0).</returns>
-    public virtual float GetOverlayImageScaleFactor(PaletteState state) => 0.5f;
+    public virtual float GetOverlayImageScaleFactor(PaletteState state) => _overlayImage.ScaleFactor;
 
     /// <summary>
     /// Gets the fixed size for the overlay image (used when scale mode is FixedSize).
     /// </summary>
     /// <param name="state">The state for which the overlay fixed size is needed.</param>
     /// <returns>Fixed size.</returns>
-    public virtual Size GetOverlayImageFixedSize(PaletteState state) => new Size(16, 16);
+    public virtual Size GetOverlayImageFixedSize(PaletteState state) => _overlayImage.FixedSize;
 
     #endregion
 }